dax-optimizer-streamlit-v1.1/requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.1.4
pyyaml==6.0.1
streamlit-extras>=0.3.0
streamlit-lottie>=0.0.5
requests>=2.27.0
